feat: Move sink_manager to core/managers/
- Move src/fapilog/_internal/sink_manager.py to src/fapilog/core/managers/
- Update import statements in:
  - src/fapilog/_internal/component_factory.py
  - src/fapilog/container.py
  - tests/test_component_factory.py
  - tests/test_sink_manager.py
  - tests/test_container_easy_wins.py
- Fix relative imports within sink_manager.py
- Update all patch paths in test files (26 references updated)

✅ SinkManager successfully moved to new location (3/12 files moved)
