+++ src/fapilog/_internal/component_factory.py
@@ -11,9 +11,9 @@ from fapilog._internal.async_lock_manager import ProcessorLockManager
 from fapilog._internal.metrics import MetricsCollector
 from fapilog._internal.middleware_manager import MiddlewareManager
 from fapilog._internal.processor_metrics import ProcessorMetrics
-from fapilog._internal.sink_manager import SinkManager
 from fapilog.core.managers.configuration_manager import ConfigurationManager
 from fapilog.core.managers.lifecycle_manager import LifecycleManager
+from fapilog.core.managers.sink_manager import SinkManager
 from fapilog.enrichers import (
     AsyncSmartCache,
     EnricherErrorHandler,
